requirements.txt
auto_mix_prep==0.2.0
Flask==2.3.2
Flask_Cors==4.0.0
flask_restx==1.1.0
psycopg2_binary==2.9.6
python-dotenv==1.0.0
